JaStockEdit リファクタリング
